Final App for Heroku
Fixed error with handling number
Finalized app wording and directing

When matt was playing around with the syntactic sugar, he had the TO phone
number passed as a hash, rather than a string.  Had to figure that out.
Done.
Then updated the app wording, added some page links (so it was easy to
redirect to the "New Message" page).

Still need to make it pretty with some styling and stuff, but the basic
structure is there and what is most important.